Add support for Netbox 3.x
This commit add support for Netbox v3.x. Maintains support for Netbox v2.11.
Core functionality is untouched, however owing to the addition of a
`get_absolute_url` method for the PDUConfig class, behaviour when adding a
single PDUConfig is changed slightly. Only difference is upon successful
addition of the PDU Config you will be redirected to the device_type for
which the PDUConfig applies, rather than back to the PDUConfig list page.

This includes the following changes:
  - Add a get_absolute_url method of PDUConfig class
  - Add 3.x templates to use updated 3.x themes and icons.
  - Add code to use the appropriate templates based on running version